Cleanup services file
